Introduce automatic caching of public keys
This default implementation uses ActiveSupport::Cache::MemoryStore,
which I think will work well in most scenarios. While it will not
persist across restarts or web server processes, it is _much_ faster
than reading from memcached or redis and it is assumed that the key
server itself is fairly fast. Therefore, it's not a big deal to miss the
cache more often than a more persistent store in favor of a faster
store.