Verify `aud` claim by default
The `aud` claim must include the full path of the original request
(without the query string).

This will prevent a JWT from being used for anything other than its
intended purpose.

It is also possible to opt out of verifying `aud` by defining
`jwt_options_for_verification` in the controller.

For example, to verify the audience is simply the name of this app, try
something like this:

    def jwt_options_for_verification
      { verify_aud: true, aud: ENV.fetch('APP_NAME') }
    end

Or, to disable this type of verification entirely:

    def jwt_options_for_verification
      {}
    end

It is not recommended to disable this verification unless some other
method is used to authorize the data returned (such as using something
in the token to scope the data returned).