Fix controller test helpers
v0.5.0 introduced changes to how automatic key selection works. In
particular, a top-level `debug_jwt` method was added in order to give
access to the payload of a JWT without verifying the signature. This
needs to be stubbed in the test helpers to prevent lots of errors in the
default tests.